Smaller transposition table size

+++ pkg/evaluation/base_material.go
@@ -6,12 +6,12 @@ import (
 )
 
 var (
-	PieceValue = [types.PIECE_TYPE_NUMBER]int{100, 300, 300, 500, 800, 2000}
+	PieceValue = [types.PIECE_TYPE_NUMBER]int16{100, 300, 300, 500, 800, 2000}
 )
 
 func (e *eval) evalBaseMaterial(pos *position.Position) {
 	// Basic Material Score
 	for pieceType := types.PAWN; pieceType < types.PIECE_TYPE_NUMBER; pieceType++ {
-		e.baseScore += PieceValue[pieceType] * (pos.PiecesBitboard[types.WHITE][pieceType].PopulationCount() - pos.PiecesBitboard[types.BLACK][pieceType].PopulationCount())
+		e.baseScore += PieceValue[pieceType] * int16(pos.PiecesBitboard[types.WHITE][pieceType].PopulationCount()-pos.PiecesBitboard[types.BLACK][pieceType].PopulationCount())
 	}
 }
